requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
chardet==4.0.0
click==8.1.3
colorama==0.4.5
Django==3.2.15
djangorestframework==3.13.1
idna==2.10
importlib-metadata==4.12.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pymemcache==3.5.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.3.0
urllib3==1.26.12
zipp==3.8.1
